ui-tests/requirements.txt
pytest>=7.4.0
selenium>=4.15.0
allure-pytest>=2.13.0
webdriver-manager>=4.0.0
pytest-xdist>=3.5.0
pytest-rerunfailures>=13.0
pytest-html>=4.1.0
Faker>=22.0.0
